RavenDB-18885 - Unskip remaining ShardedAtomicClusterReadWriteTests
